+++ test_conformance/subgroups/main.cpp
@@ -30,50 +30,17 @@ test_definition test_list[] = {
     ADD_TEST_VERSION(work_group_functions_ext, Version(2, 0)),
     ADD_TEST_VERSION(work_group_functions_core, Version(2, 1)),
     ADD_TEST_VERSION(barrier_functions_ext, Version(2, 0)),
-    ADD_TEST_VERSION(barrier_functions_core, Version(2, 1))
+    ADD_TEST_VERSION(barrier_functions_core, Version(2, 1)),
+    ADD_TEST_VERSION(ifp_ext, Version(2, 0)),
+    ADD_TEST_VERSION(ifp_core, Version(2, 1))
 };
 
 const int test_num = ARRAY_SIZE(test_list);
-bool gUseCoreSubgroups = false;
-bool gTestIFP = true;
-
-static test_status checkIFPSupport(cl_device_id device)
-{
-    cl_uint ifp_supported;
-    cl_uint error;
-    error = clGetDeviceInfo(device,
-                            CL_DEVICE_SUB_GROUP_INDEPENDENT_FORWARD_PROGRESS,
-                            sizeof(ifp_supported), &ifp_supported, NULL);
-    if (error != CL_SUCCESS)
-    {
-        print_error(
-            error,
-            "Unable to get CL_DEVICE_SUB_GROUP_INDEPENDENT_FORWARD_PROGRESS "
-            "capability");
-        return TEST_FAIL;
-    }
-    // skip testing ifp
-    if (ifp_supported != 1)
-    {
-        log_info("INDEPENDENT FORWARD PROGRESS not supported...\n");
-        gTestIFP = false;
-    }
-    else
-    {
-        log_info("INDEPENDENT FORWARD PROGRESS supported...\n");
-    }
-    return TEST_PASS;
-}
 
 static test_status InitCL(cl_device_id device)
 {
     auto version = get_device_cl_version(device);
     test_status ret = TEST_PASS;
-    ret = checkIFPSupport(device);
-    if (ret != TEST_PASS)
-    {
-        return ret;
-    }
     if (version >= Version(3, 0))
     {
         cl_uint max_sub_groups;
