+++ test_conformance/subgroups/procs.h
@@ -26,8 +26,6 @@
 
 extern MTdata gMTdata;
 
-extern bool gUseCoreSubgroups;
-extern bool gTestIFP;
 extern int test_sub_group_info_ext(cl_device_id device, cl_context context,
                                    cl_command_queue queue, int num_elements);
 extern int test_sub_group_info_core(cl_device_id device, cl_context context,
@@ -54,5 +52,9 @@ extern int test_barrier_functions_core(cl_device_id device, cl_context context,
                                        int num_elements);
 extern int test_pipe_functions(cl_device_id device, cl_context context,
                                cl_command_queue queue, int num_elements);
+extern int test_ifp_ext(cl_device_id device, cl_context context,
+                        cl_command_queue queue, int num_elements);
+extern int test_ifp_core(cl_device_id device, cl_context context,
+                         cl_command_queue queue, int num_elements);
 
 #endif /*_procs_h*/
