+++ test_conformance/subgroups/test_queries.cpp
@@ -14,6 +14,7 @@
 // limitations under the License.
 //
 #include "procs.h"
+#include "subhelpers.h"
 
 typedef struct
 {
@@ -23,7 +24,8 @@ typedef struct
 
 
 int test_sub_group_info(cl_device_id device, cl_context context,
-                        cl_command_queue queue, int num_elements)
+                        cl_command_queue queue, int num_elements,
+                        bool useCoreSubgroups)
 {
     static const size_t gsize0 = 80;
     int i, error;
@@ -44,28 +46,29 @@ int test_sub_group_info(cl_device_id device, cl_context context,
     clProgramWrapper program;
     clKernelWrapper kernel;
     clMemWrapper out;
+    std::stringstream kernel_sstr;
+    if (useCoreSubgroups)
+    {
+        kernel_sstr << "#pragma OPENCL EXTENSION cl_khr_subgroups : enable\n";
+    }
+    kernel_sstr
+        << "\n"
+           "typedef struct {\n"
+           "    uint maxSubGroupSize;\n"
+           "    uint numSubGroups;\n"
+           "} result_data;\n"
+           "\n"
+           "__kernel void query_kernel( __global result_data *outData )\n"
+           "{\n"
+           "    int gid = get_global_id( 0 );\n"
+           "    outData[gid].maxSubGroupSize = get_max_sub_group_size();\n"
+           "    outData[gid].numSubGroups = get_num_sub_groups();\n"
+           "}";
 
-    std::string pragma_str = gUseCoreSubgroups
-        ? "\n"
-        : "#pragma OPENCL EXTENSION cl_khr_subgroups : enable\n";
-    std::string kernel_str = pragma_str
-        + "\n"
-          "typedef struct {\n"
-          "    uint maxSubGroupSize;\n"
-          "    uint numSubGroups;\n"
-          "} result_data;\n"
-          "\n"
-          "__kernel void query_kernel( __global result_data *outData )\n"
-          "{\n"
-          "    int gid = get_global_id( 0 );\n"
-          "    outData[gid].maxSubGroupSize = get_max_sub_group_size();\n"
-          "    outData[gid].numSubGroups = get_num_sub_groups();\n"
-          "}";
-
-    const char *query_kernel_source = kernel_str.c_str();
-
+    const std::string &kernel_str = kernel_sstr.str();
+    const char *kernel_src = kernel_str.c_str();
     error = create_single_kernel_helper_with_build_options(
-        context, &program, &kernel, 1, &query_kernel_source, "query_kernel",
+        context, &program, &kernel, 1, &kernel_src, "query_kernel",
         "-cl-std=CL2.0");
     if (error != 0) return error;
 
@@ -93,94 +96,51 @@ int test_sub_group_info(cl_device_id device, cl_context context,
                             (void *)&platform, NULL);
     test_error(error, "clDeviceInfo failed for CL_DEVICE_PLATFORM");
 
-    if (gUseCoreSubgroups)
+    subgroupsAPI subgroupsApiSet(platform, useCoreSubgroups);
+    clGetKernelSubGroupInfoKHR_fn f_ptr = subgroupsApiSet.get_f_ptr();
+    if (f_ptr == NULL)
     {
-        error = clGetKernelSubGroupInfo(
-            kernel, device, CL_KERNEL_MAX_SUB_GROUP_SIZE_FOR_NDRANGE,
-            sizeof(local), (void *)&local, sizeof(kernel_max_subgroup_size),
-            (void *)&kernel_max_subgroup_size, &realSize);
-        test_error(error,
-                   "clGetKernelSubGroupInfo failed for "
-                   "CL_KERNEL_MAX_SUB_GROUP_SIZE_FOR_NDRANGE");
-        log_info("The CL_KERNEL_MAX_SUB_GROUP_SIZE_FOR_NDRANGE for the kernel "
-                 "is %d.\n",
-                 (int)kernel_max_subgroup_size);
-        if (realSize != sizeof(kernel_max_subgroup_size))
-        {
-            log_error("ERROR: Returned size of max sub group size not valid! "
-                      "(Expected %d, got %d)\n",
-                      (int)sizeof(kernel_max_subgroup_size), (int)realSize);
-            return -1;
-        }
-        error = clGetKernelSubGroupInfo(
-            kernel, device, CL_KERNEL_SUB_GROUP_COUNT_FOR_NDRANGE,
-            sizeof(local), (void *)&local, sizeof(kernel_subgroup_count),
-            (void *)&kernel_subgroup_count, &realSize);
-        test_error(error,
-                   "clGetKernelSubGroupInfo failed for "
-                   "CL_KERNEL_SUB_GROUP_COUNT_FOR_NDRANGE");
-        log_info(
-            "The CL_KERNEL_SUB_GROUP_COUNT_FOR_NDRANGE for the kernel is %d.\n",
-            (int)kernel_subgroup_count);
-
-        if (realSize != sizeof(kernel_subgroup_count))
-        {
-            log_error("ERROR: Returned size of sub group count not valid! "
-                      "(Expected %d, got %d)\n",
-                      (int)sizeof(kernel_subgroup_count), (int)realSize);
-            return -1;
-        }
+        log_error("ERROR: %s function not available", subgroupsApiSet.f_name);
+        return TEST_FAIL;
     }
-    else
+
+    error =
+        f_ptr(kernel, device, subgroupsApiSet.get_enum_max_size(),
+              sizeof(local), (void *)&local, sizeof(kernel_max_subgroup_size),
+              (void *)&kernel_max_subgroup_size, &realSize);
+    if (error != CL_SUCCESS)
     {
-        clGetKernelSubGroupInfoKHR_fn clGetKernelSubGroupInfoKHR_ptr;
-        clGetKernelSubGroupInfoKHR_ptr = (clGetKernelSubGroupInfoKHR_fn)
-            clGetExtensionFunctionAddressForPlatform(
-                platform, "clGetKernelSubGroupInfoKHR");
-        if (clGetKernelSubGroupInfoKHR_ptr == NULL)
-        {
-            log_error(
-                "ERROR: clGetKernelSubGroupInfoKHR function not available");
-            return -1;
-        }
+        log_error("ERROR: %s function error for %s", subgroupsApiSet.f_name,
+                  subgroupsApiSet.enum_max_size_name);
+        return TEST_FAIL;
+    }
+    log_info("The %s for the kernel is %d.\n",
+             subgroupsApiSet.enum_max_size_name, (int)kernel_max_subgroup_size);
+    if (realSize != sizeof(kernel_max_subgroup_size))
+    {
+        log_error("ERROR: Returned size of max sub group size not valid! "
+                  "(Expected %d, got %d)\n",
+                  (int)sizeof(kernel_max_subgroup_size), (int)realSize);
+        return TEST_FAIL;
+    }
+    error = f_ptr(kernel, device, subgroupsApiSet.get_enum_count(),
+                  sizeof(local), (void *)&local, sizeof(kernel_subgroup_count),
+                  (void *)&kernel_subgroup_count, &realSize);
+    if (error != CL_SUCCESS)
+    {
+        log_error("ERROR: %s function error for %s", subgroupsApiSet.f_name,
+                  subgroupsApiSet.enum_count_name);
+        return TEST_FAIL;
+    }
+    log_info("The %s for the kernel is %d.\n", subgroupsApiSet.enum_count_name,
+             (int)kernel_subgroup_count);
 
-        error = clGetKernelSubGroupInfoKHR_ptr(
-            kernel, device, CL_KERNEL_MAX_SUB_GROUP_SIZE_FOR_NDRANGE_KHR,
-            sizeof(local), (void *)&local, sizeof(kernel_max_subgroup_size),
-            (void *)&kernel_max_subgroup_size, &realSize);
-        test_error(error,
-                   "clGetKernelSubGroupInfoKHR failed for "
-                   "CL_KERNEL_MAX_SUB_GROUP_SIZE_FOR_NDRANGE_KHR");
-        log_info(
-            "The CL_KERNEL_MAX_SUB_GROUP_SIZE_FOR_NDRANGE_KHR for the kernel "
-            "is %d.\n",
-            (int)kernel_max_subgroup_size);
-        if (realSize != sizeof(kernel_max_subgroup_size))
-        {
-            log_error("ERROR: Returned size of max sub group size not valid! "
-                      "(Expected %d, got %d)\n",
-                      (int)sizeof(kernel_max_subgroup_size), (int)realSize);
-            return -1;
-        }
-
-        error = clGetKernelSubGroupInfoKHR_ptr(
-            kernel, device, CL_KERNEL_SUB_GROUP_COUNT_FOR_NDRANGE_KHR,
-            sizeof(local), (void *)&local, sizeof(kernel_subgroup_count),
-            (void *)&kernel_subgroup_count, &realSize);
-        test_error(error,
-                   "clGetKernelSubGroupInfoKHR failed for "
-                   "CL_KERNEL_SUB_GROUP_COUNT_FOR_NDRANGE_KHR");
-        log_info("The CL_KERNEL_SUB_GROUP_COUNT_FOR_NDRANGE_KHR for the kernel "
-                 "is %d.\n",
-                 (int)kernel_subgroup_count);
-
-        if (realSize != sizeof(kernel_subgroup_count))
-        {
-            log_error("ERROR: Returned size of sub group count not valid! "
-                      "(Expected %d, got %d)\n",
-                      (int)sizeof(kernel_subgroup_count), (int)realSize);
-            return -1;
-        }
+    if (realSize != sizeof(kernel_subgroup_count))
+    {
+        log_error("ERROR: Returned size of sub group count not valid! "
+                  "(Expected %d, got %d)\n",
+                  (int)sizeof(kernel_subgroup_count), (int)realSize);
+        return TEST_FAIL;
     }
 
     // Verify that the kernel gets the same max_subgroup_size and subgroup_count
@@ -227,14 +187,12 @@ int test_sub_group_info(cl_device_id device, cl_context context,
 int test_sub_group_info_core(cl_device_id device, cl_context context,
                              cl_command_queue queue, int num_elements)
 {
-    gUseCoreSubgroups = true;
-    return test_sub_group_info(device, context, queue, num_elements);
+    return test_sub_group_info(device, context, queue, num_elements, true);
 }
 
 int test_sub_group_info_ext(cl_device_id device, cl_context context,
                             cl_command_queue queue, int num_elements)
 {
-    gUseCoreSubgroups = false;
     bool hasExtension = is_extension_available(device, "cl_khr_subgroups");
 
     if (!hasExtension)
@@ -244,5 +202,5 @@ int test_sub_group_info_ext(cl_device_id device, cl_context context,
         return TEST_SKIP;
     }
 
-    return test_sub_group_info(device, context, queue, num_elements);
+    return test_sub_group_info(device, context, queue, num_elements, false);
 }
