+++ test_conformance/subgroups/test_workitem.cpp
@@ -184,7 +184,8 @@ static int check_group(const get_test_data *result, int nw, cl_uint ensg,
 }
 
 int test_work_item_functions(cl_device_id device, cl_context context,
-                             cl_command_queue queue, int num_elements)
+                             cl_command_queue queue, int num_elements,
+                             bool useCoreSubgroups)
 {
     static const size_t lsize = 200;
     int error;
@@ -197,35 +198,37 @@ int test_work_item_functions(cl_device_id device, cl_context context,
     clProgramWrapper program;
     clKernelWrapper kernel;
     clMemWrapper out;
-
-    std::string pragma_str = gUseCoreSubgroups
-        ? "\n"
-        : "#pragma OPENCL EXTENSION cl_khr_subgroups : enable\n";
-    std::string kernel_str = pragma_str
-        + "\n"
-          "\n"
-          "typedef struct {\n"
-          "    uint subGroupSize;\n"
-          "    uint maxSubGroupSize;\n"
-          "    uint numSubGroups;\n"
-          "    uint enqNumSubGroups;\n"
-          "    uint subGroupId;\n"
-          "    uint subGroupLocalId;\n"
-          "} get_test_data;\n"
-          "\n"
-          "__kernel void get_test( __global get_test_data *outData )\n"
-          "{\n"
-          "    int gid = get_global_id( 0 );\n"
-          "    outData[gid].subGroupSize = get_sub_group_size();\n"
-          "    outData[gid].maxSubGroupSize = get_max_sub_group_size();\n"
-          "    outData[gid].numSubGroups = get_num_sub_groups();\n"
-          "    outData[gid].enqNumSubGroups = get_enqueued_num_sub_groups();\n"
-          "    outData[gid].subGroupId = get_sub_group_id();\n"
-          "    outData[gid].subGroupLocalId = get_sub_group_local_id();\n"
-          "}";
-    const char *get_test_source = kernel_str.c_str();
+    std::stringstream kernel_sstr;
+    if (useCoreSubgroups)
+    {
+        kernel_sstr << "#pragma OPENCL EXTENSION cl_khr_subgroups : enable\n";
+    }
+    kernel_sstr
+        << "\n"
+           "\n"
+           "typedef struct {\n"
+           "    uint subGroupSize;\n"
+           "    uint maxSubGroupSize;\n"
+           "    uint numSubGroups;\n"
+           "    uint enqNumSubGroups;\n"
+           "    uint subGroupId;\n"
+           "    uint subGroupLocalId;\n"
+           "} get_test_data;\n"
+           "\n"
+           "__kernel void get_test( __global get_test_data *outData )\n"
+           "{\n"
+           "    int gid = get_global_id( 0 );\n"
+           "    outData[gid].subGroupSize = get_sub_group_size();\n"
+           "    outData[gid].maxSubGroupSize = get_max_sub_group_size();\n"
+           "    outData[gid].numSubGroups = get_num_sub_groups();\n"
+           "    outData[gid].enqNumSubGroups = get_enqueued_num_sub_groups();\n"
+           "    outData[gid].subGroupId = get_sub_group_id();\n"
+           "    outData[gid].subGroupLocalId = get_sub_group_local_id();\n"
+           "}";
+    const std::string &kernel_str = kernel_sstr.str();
+    const char *kernel_src = kernel_str.c_str();
     error = create_single_kernel_helper_with_build_options(
-        context, &program, &kernel, 1, &get_test_source, "get_test",
+        context, &program, &kernel, 1, &kernel_src, "get_test",
         "-cl-std=CL2.0");
     if (error != 0) return error;
 
@@ -308,14 +311,12 @@ int test_work_item_functions(cl_device_id device, cl_context context,
 int test_work_item_functions_core(cl_device_id device, cl_context context,
                                   cl_command_queue queue, int num_elements)
 {
-    gUseCoreSubgroups = true;
-    return test_work_item_functions(device, context, queue, num_elements);
+    return test_work_item_functions(device, context, queue, num_elements, true);
 }
 
 int test_work_item_functions_ext(cl_device_id device, cl_context context,
                                  cl_command_queue queue, int num_elements)
 {
-    gUseCoreSubgroups = false;
     bool hasExtension = is_extension_available(device, "cl_khr_subgroups");
 
     if (!hasExtension)
@@ -325,5 +326,6 @@ int test_work_item_functions_ext(cl_device_id device, cl_context context,
         return TEST_SKIP;
     }
 
-    return test_work_item_functions(device, context, queue, num_elements);
+    return test_work_item_functions(device, context, queue, num_elements,
+                                    false);
 }
